requirements.txt
django==4.2.5
djangorestframework==3.14.0
pandas==2.1.1
scikit-learn==1.3.1
nltk==3.8.1